refactor: move code
